+++ metadata/apps/example-hasura/hasura-request-logger.js
@@ -1,0 +1,14 @@
+// import 'colors'; // This is known issue with this package...
+
+const hasuraRequestLogger = (req, res, next) => {
+  console.log(`[${req.originalUrl}] Request hit!`);
+  // --------------------------------------------------------------------------------
+  // Get variables
+  console.log('This is what we received from Hasura when calling the hook');
+  console.log(req.body);
+  console.log(Array(80 + 1).join('-'));
+  // --------------------------------------------------------------------------------
+  next();
+};
+
+export default hasuraRequestLogger;