+++ src/routes/metadata-management/router.metadata-import.js
@@ -39,12 +39,16 @@ router.post(
 
     // Write file in data
     await fsPromise
-      .writeFile(path.resolve(process.cwd(), './data/last-metadata-import-timestamp'), `${Date.now()}`, 'utf8')
+      .writeFile(
+        path.resolve(process.cwd(), './data/metadata-history/last-metadata-import-timestamp'),
+        `${Date.now()}`,
+        'utf8'
+      )
       .then((data) => {
         return data;
       })
       .catch((error) => {
-        console.error("[restart-microservice] ERROR: can't write ./data/last-metadata-import-timestamp");
+        console.error("[metadata-import] ERROR: can't write ./data/metadata-history/last-metadata-import-timestamp");
         console.error(error);
       });